# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação API backend",
        "description": "Documentação API backend - CRUD",
        "contact": {
            "email": "assuncaofilipe97@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/cake/{id}/subscribe": {
            "post": {
                "tags": [
                    "subscribe"
                ],
                "summary": "Receber atualização de disponibilidade de um bolo",
                "description": "É realizado a inscrição no bolo desejado para receber um email caso o bolo fique disponível.",
                "operationId": "6247c97c4441132a36f16867e4abded5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "maxLength": 150
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "change-status": {
                                    "summary": "Fazer inscrição de atualizações de um bolo",
                                    "value": {
                                        "email": "zezinho@gmail.com"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inscrição de atualizações.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": "true",
                                    "data": {
                                        "message": "Inscrição realizada com sucesso."
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "email": {
                                                    "type": "string"
                                                },
                                                "password": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                },
                                "examples": {
                                    "obrigatório": {
                                        "summary": "Campos obrigatórios",
                                        "value": {
                                            "success": "false",
                                            "data": {
                                                "email": [
                                                    "O campo email é obrigatório."
                                                ]
                                            }
                                        }
                                    },
                                    "em uso": {
                                        "summary": "Em uso",
                                        "value": {
                                            "success": "false",
                                            "data": {
                                                "email": [
                                                    "O campo email já está sendo utilizado."
                                                ]
                                            }
                                        }
                                    },
                                    "Inválido": {
                                        "summary": "Inválido",
                                        "value": {
                                            "success": "false",
                                            "data": {
                                                "email": [
                                                    "O campo email deve ser um endereço de e-mail válido."
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}